+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ---
 
+#### 14.07.2018
+
+* [Rust] Code of Kutulu - Initial file (CodinGame)
+
 #### 13.07.2018
 
 * [Rust] CodeRoyale - General refactoring adujustments (CodinGame)